# Perf improvements to `scheduler`