--- a/dotcom-rendering/package.json
+++ b/dotcom-rendering/package.json
@@ -74,7 +74,7 @@
 		"@guardian/eslint-plugin-source-react-components": "18.0.0",
 		"@guardian/identity-auth": "1.1.0",
 		"@guardian/identity-auth-frontend": "1.0.0",
-		"@guardian/libs": "15.7.0",
+		"@guardian/libs": "15.7.1",
 		"@guardian/prettier": "5.0.0",
 		"@guardian/shimport": "1.0.2",
 		"@guardian/source-foundations": "13.0.0",