<!-- In this repo you can label a PR with the "PR Deployment" label to deploy the code to a publicly accessible url -->

## What does this change?

- bumps libs to take advantage of [more efficient logging](https://github.com/guardian/csnx/pull/809)
- refactors `scheduler` to make it faster

## Why?

Preliminary numbers showed the scheduler seemed to slow things down too much, even though it was effectively working in passthru mode.

We need to get it closer to negligible effect before we can start scheduling things.

## Difference

Using [tachometer](https://www.npmjs.com/package/tachometer), running 1000 tasks that simply return `Promise.resolve()` on my work M1 Max mbp, I get these results:

| Unscheduled (baseline) | Previously | Now | 
| ----------- | ---------- | ---------- |
| ~0.25ms | ~20.14ms | ~4.6ms |

it drops to ~1.75ms if we don't measure the duration of tasks, but that slightly defeats the purpose...

## Failed attempts
I tried using https://www.npmjs.com/package/flatqueue instead of the naive one this scheduler uses, but it had no effect until running 10k tasks, and then only slightly.